replace std::size_t to std::uint32_t

+++ include/Algorithms/GeneticAlgorithm.hpp
@@ -12,11 +12,11 @@ namespace algorithm
 class GeneticAlgorithm final 
     : public AlgorithmBase<
             utils::GAExecutionResult, 
-            std::size_t,
-            std::size_t,
-            std::size_t,
-            std::size_t,
-            std::size_t,
+            std::uint32_t,
+            std::uint32_t,
+            std::uint32_t,
+            std::uint32_t,
+            std::uint32_t,
             utils::selection_function   
         >
 {
@@ -34,11 +34,11 @@ public:
     ~GeneticAlgorithm() = default;
 
     utils::GAExecutionResult Execute(
-        const std::size_t kIterations,
-        const std::size_t kPopulation,
-        const std::size_t kCrossovers,
-        const std::size_t kMutations,
-        const std::size_t kAmountGensMutation,
+        const std::uint32_t kIterations,
+        const std::uint32_t kPopulation,
+        const std::uint32_t kCrossovers,
+        const std::uint32_t kMutations,
+        const std::uint32_t kAmountGensMutation,
         utils::selection_function sf
     ) override;
 
@@ -46,22 +46,22 @@ private:
 
     void Crossover(
         std::vector<model::Candidate>& candidates,
-        const std::size_t kCrossovers,
-        const std::size_t kPopulation,
+        const std::uint32_t kCrossovers,
+        const std::uint32_t kPopulation,
         utils::selection_function sf
     );
 
     void Mutate(
         std::vector<model::Candidate>& candidates,
-        const std::size_t kPopulation,
-        const std::size_t kMutations,
-        const std::size_t kAmountGensMutation,
+        const std::uint32_t kPopulation,
+        const std::uint32_t kMutations,
+        const std::uint32_t kAmountGensMutation,
         utils::selection_function sf
     );
 
     void DoSelection(
         std::vector<model::Candidate>& candidates,
-        const std::size_t kPopulation
+        const std::uint32_t kPopulation
     );
 
     model::Candidates candidates_;
