+++ include/Core/CNF.hpp
@@ -11,7 +11,7 @@ class CNF final
 {
 public:
     CNF() = default;
-    explicit CNF( const std::size_t kAmountVariables );
+    explicit CNF( const std::uint32_t kAmountVariables );
     ~CNF() = default;
 
     std::string& GetCNF();
@@ -19,9 +19,9 @@ public:
 
 private:
     std::string GenerateRandomCNF(
-        const std::size_t kVariableNumber, 
-        const std::size_t kClauseLength, 
-        const std::size_t kMaxClauseLength
+        const std::uint32_t kVariableNumber, 
+        const std::uint32_t kClauseLength, 
+        const std::uint32_t kMaxClauseLength
     ) const;
 
     std::string TrimAndCLean( const std::string& kString ) const;
